+++ assets/javascripts/kitten/components/form/dropdown-select/styles.js
@@ -111,7 +111,8 @@ export const StyledDropdown = styled.div`
   .k-Form-Dropdown__list {
     box-sizing: border-box;
     position: absolute;
-    z-index: 1;
+    z-index: 1000;
+    z-index: var(--menu-z-index, 1000);
     width: 100%;
     max-height: ${pxToRem(250)};
     padding: 0;